html/semantics/scripting-1/the-script-element/async_005.html

<!DOCTYPE html>
<html>
    <head>
        <title>A script element with both async and defer set should execute asynchronously</title>
        <meta name="timeout" content="long">
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures a script element with both async and defer set should execute asynchronously." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/#attr-script-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
 <div id=log></div>
        <script type="text/javascript">
        var t = async_test("A script element with both async and defer set should execute asynchronously");

        function timeout()
        {
            t.step(function(){
                var test_result = document.getElementById("testresult").innerHTML;
                // Four scripts should have executed by this point.
                assert_equals(test_result.length, 4, "Not all scripts have executed");
                // Inline script '2' should come first
                assert_equals(test_result[0], "2", "Inline script should be evaluated first, before defer async");
                // Inline script '2' should come before inline script '3' and '4'.
                assert_greater_than(test_result.indexOf("3"), test_result.indexOf("2"), "Inline script 2 executed in the incorrect order");
                assert_greater_than(test_result.indexOf("4"), test_result.indexOf("3"), "Inline script 3 executed in the incorrect order");
                assert_not_equals(test_result.indexOf("1"), -1, "Script 1 (defer async) has not executed");
            });
            t.done();
        }

        var timer = setTimeout(timeout, 5000);

        function log(text)
        {
            var textNode = document.createTextNode(text);
            document.getElementById("testresult").appendChild(textNode);
        }
        </script>

        <span id="testresult"></span>

        <script type="text/javascript" src="log.py?sec=1&id=1" defer async></script>
        <script type="text/javascript">log('2');</script>
        <script type="text/javascript" src="log.py?sec=3&id=3"></script>
        <script type="text/javascript">log('4');</script>
    </body>
</html>
