Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/async_005-expected.txt


FAIL A script element with both async and defer set should execute asynchronously assert_equals: expected "2134" but got "2341"
2341
